// src/helpers/adapters/nodeBinanceApi/interfaces/nodeBinanceApi-Interface.ts
export type SideOrderTypes = 'BUY' | 'SELL'

export type TypeOrderTypes = 'ICEBERG' | 'LIMIT' | 'MARKET' | 'STOP_LOSS' | 'STOP_LOSS_LIMIT' | 'TAKE_PROFIT' | 'TAKE_PROFIT_LIMIT'

export type StopOrderTypes = 'STOP_LOSS' | 'STOP_LOSS_LIMIT' | 'TAKE_PROFIT' | 'TAKE_PROFIT_LIMIT'

export type StatusOrderTypes = 'NEW' | 'FILLED' | 'CANCELED' | 'REJECTED'

export type _callback = (...args: any) => any

export type _asyncCallback = (...args: any) => Promise<any>

export interface SettingsInterface {
  APIKEY: string
  APISECRET: string
  urls: {
    base: string
    stream: string
  }
}

export interface OutputBalanceInterface {
  [key: string]: {
    available: string
    onOrder: string
  }
}

export interface InputNewOrderInterface {
  settings: SettingsInterface
  symbol: string
  quantity: number
  type: TypeOrderTypes
  price?: number
  options: {
    type: TypeOrderTypes
    stopPrice?: number
    icebergQuantity?: number
  }
}
// ...
export interface InputCancelInterface {
  settings: SettingsInterface
  symbol: string
  orderId: number
}
// ...
export interface InputOrderStatusInterface {
  settings: SettingsInterface
  symbol: string
  orderId: number
}

export interface OutputOrderStatusInterface {
  symbol: string
  orderId: number
  orderListId: number // Unless part of an OCO, the value will always be -1
  clientOrderId: string
  price: string
  origQty: string
  executedQty: string
  cummulativeQuoteQty: string
  status: StatusOrderTypes
  timeInForce: string
  type: TypeOrderTypes
  side: SideOrderTypes
  stopPrice: string
  icebergQty: string
  time: bigint
  updateTime: bigint
  isWorking: true
  workingTime: bigint
  origQuoteOrderQty: string
  selfTradePreventionMode: string
}

export interface InputOrderTradeInterface {
  settings: SettingsInterface
  symbol: string
  orderId: number
}

export interface OutputOrdertradeInterface {
  symbol: string
  id: number
  orderId: number
  orderListId: number
  price: string
  qty: string
  quoteQty: string
  commission: string
  commissionAsset: string
  time: bigint
  isBuyer: boolean
  isMaker: false
  isBestMatch: boolean
}
// ...
export interface InputChartStreamInterface {
  settings: SettingsInterface
  symbol: string
  interval: string
  callback: _asyncCallback
  limit?: number
}

export interface OutputOHLCInterface {
  open: number[]
  high: number[]
  low: number[]
  close: number[]
  volume: number[]
}

export interface NodeBinanceApiAdapterInterface {
  exchangeInfo: (settings: SettingsInterface) => Promise<any>
  exchangeBalance: (settings: SettingsInterface) => Promise<any>
  buy: (data: InputNewOrderInterface) => Promise<any>
  sell: (data: InputNewOrderInterface) => Promise<any>
  cancel: (data: InputCancelInterface) => Promise<any>
  orderTrade: (data: InputOrderTradeInterface) => Promise<OutputOrdertradeInterface>
  orderStatus: (data: InputOrderStatusInterface) => Promise<OutputOrderStatusInterface>
  miniTickerStream: (callback: _callback, settings: SettingsInterface) => Promise<void>
  chartStream: (data: InputChartStreamInterface) => Promise<void>
  userDataStream: (callback: _callback, executionCallback: boolean, listStatusCallback: _callback, settings: SettingsInterface) => Promise<void>
}